Improve env endpoint section
While we're talking about the format of the env endpoint, I think it
would be worthwhile to mention the `env/{id}` format that provides the
value used in the environment and which property source contributed it.

The property sources are listed in order of precedence, which is handy
when figuring out why a given key is behaving differently than expected.